+++ lib/pwnlib/shellcraft/generators/amd64/common/infloop.rb
