+++ lib/pwnlib/shellcraft/generators/amd64/common/nop.rb
