+++ lib/pwnlib/shellcraft/generators/amd64/common/popad.rb
