+++ lib/pwnlib/shellcraft/generators/amd64/common/pushstr.rb
