+++ lib/pwnlib/shellcraft/generators/amd64/common/pushstr_array.rb
