+++ lib/pwnlib/shellcraft/generators/amd64/common/setregs.rb
