+++ lib/pwnlib/shellcraft/generators/amd64/linux/syscall.rb
