+++ lib/pwnlib/shellcraft/generators/i386/common/infloop.rb
