+++ lib/pwnlib/shellcraft/generators/i386/common/pushstr_array.rb
