+++ lib/pwnlib/shellcraft/generators/i386/common/setregs.rb
