+++ lib/pwnlib/shellcraft/generators/i386/linux/syscall.rb
