--- lib/pwnlib/shellcraft/generators/i386/nop.rb
@@ -1,5 +1,0 @@
-# encoding: ASCII-8BIT
-
-::Pwnlib::Shellcraft.define(__FILE__) do
-  cat 'nop'
-end
